Update the sample to MSAL 3.x and ADAL 5.x
