--- TokenCacheMigration/CommonCacheAdalV4/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.IdentityModel.Clients.ActiveDirectory" version="4.4.0" allowedVersions="[4.1.0,5)" targetFramework="net461" />
-</packages>
