+++ TokenCacheMigration/CommonCacheAdalV5/Properties/AssemblyInfo.cs
