+++ TokenCacheMigration/CommonCacheAdalV5/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.IdentityModel.Clients.ActiveDirectory" version="5.0.2-preview" allowedVersions="[4.1.0,6)" targetFramework="net461" />
+</packages>
