+++ TokenCacheMigration/CommonCacheAdalv3/Program.cs
@@ -16,7 +16,6 @@ namespace CommonCacheADAL
         {
             Console.WriteLine(Path.GetFileNameWithoutExtension(Assembly.GetEntryAssembly().Location));
             DoIt().Wait();
-            Console.ReadLine();
         }
 
         static async Task DoIt()
