--- TokenCacheMigration/CommonCacheMsal2/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.Identity.Client" version="2.4.0-preview" allowedVersions="[2,3)" targetFramework="net461" />
-</packages>
