+++ TokenCacheMigration/CommonCacheMsalV3/App.config
