+++ TokenCacheMigration/CommonCacheMsalV3/Properties/AssemblyInfo.cs
