+++ TokenCacheMigration/CommonCacheMsalV3/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.Identity.Client" version="3.0.4-preview" targetFramework="net461" />
+</packages>
