+++ TokenCacheMigration/TestAll.bat
@@ -1,31 +1,31 @@
 del cacheAdalV3.bin
 del unifiedCache.bin
-CommonCacheAdalv3\bin\Debug\CommonCacheAdalV3.exe
-CommonCacheAdalV4\bin\Debug\CommonCacheAdalV4.exe
+CommonCacheAdalV3\bin\Debug\CommonCacheAdalV3.exe
+CommonCacheAdalV5\bin\Debug\CommonCacheAdalV5.exe
 
 del cacheAdalV3.bin
 del unifiedCache.bin
-CommonCacheAdalv3\bin\Debug\CommonCacheAdalV3.exe
-CommonCacheMsal2\bin\Debug\CommonCacheMsal2.exe
+CommonCacheAdalV3\bin\Debug\CommonCacheAdalV3.exe
+CommonCacheMsalV3\bin\Debug\CommonCacheMsalV3.exe
 
 del cacheAdalV3.bin
 del unifiedCache.bin
-CommonCacheAdalV4\bin\Debug\CommonCacheAdalV4.exe
-CommonCacheAdalv3\bin\Debug\CommonCacheAdalV3.exe
+CommonCacheAdalV5\bin\Debug\CommonCacheAdalV5.exe
+CommonCacheAdalV3\bin\Debug\CommonCacheAdalV3.exe
 
 del cacheAdalV3.bin
 del unifiedCache.bin
-CommonCacheAdalV4\bin\Debug\CommonCacheAdalV4.exe
-CommonCacheMsal2\bin\Debug\CommonCacheMsal2.exe
+CommonCacheAdalV5\bin\Debug\CommonCacheAdalV5.exe
+CommonCacheMsalV3\bin\Debug\CommonCacheMsalV3.exe
 
 del cacheAdalV3.bin
 del unifiedCache.bin
-CommonCacheMsal2\bin\Debug\CommonCacheMsal2.exe
-CommonCacheAdalv3\bin\Debug\CommonCacheAdalV3.exe
+CommonCacheMsalV3\bin\Debug\CommonCacheMsalV3.exe
+CommonCacheAdalV3\bin\Debug\CommonCacheAdalV3.exe
 
 del cacheAdalV3.bin
 del unifiedCache.bin
-CommonCacheMsal2\bin\Debug\CommonCacheMsal2.exe
-CommonCacheAdalV4\bin\Debug\CommonCacheAdalV4.exe
+CommonCacheMsalV3\bin\Debug\CommonCacheMsalV3.exe
+CommonCacheAdalV5\bin\Debug\CommonCacheAdalV5.exe
 
 pause
